Decouple walking action from the player class
The code for walking is now stored in its own separate file called
WalkAction. This way it can be used for multiple different entities and
doesn't have to be hardcoded in the player's class.

A newly created Action is not assigned to any entity upon creation. The
bind(entity) function has to be used to assign it.