Prepare code for turn based system
The player is stored in a table which will later contain all actors. The
draw function loops over all actors and calls their respective draw
function.

The update function is where the turn based magic will happen. It is
basically a "resumable" loop which iterates over all actors. It gets the
next action stored for the actor it currently looks at and performs that
action, jumping to the next actor in the table.

If an actor doesn't have an action (this will only ever happen with the
player's character) the loop stops and does nothing until a valid action
is input.

It will proceed until it reaches the end of the table, then jumps back
to the beginning of the table and continues until it encounters the
player again.