Implement energy based actions per turn
Each turn all actors are granted a certain amount of energy. If their
energy is equal to, or surpasses the threshold which is needed to
perform an action they perform the action and their energy level is
reset.

The player always gets enough energy to make sure he can perform one
action per turn. The values of all other characters are relative to
this value.