Use raycasting to simulate a FOV for the player character
It casts rays with a slightly randomized length in a 360° radius around
the player's tile. When the ray hits a tile, the visibility flag is set
to true. If the tile is impassable the ray stops and therefore makes
everything beyond that wall invisible to the player.

Tiles which are not visible to the player are rendered with a dark grey
and actors outside of the FOV are hidden completely.