Expand the combat system
The combat system is still kept simple, but I added attack ratings and
defense ratings for each actor (currently as hardcoded values). These
values are used to determine if an attacker hits the defender.

The bigger the AR is, the better the chance of a successful hit will be.
The difference between the AR and DR will determine what value the
attacker will have to roll on a 3d6 to make a hit.

Even if the AR is really low compared to the DR there is a small chance
that the attacker will score a hit.

Actors now have a health attribute which will be drained by attacks. The
"Wait" action restores a small amount per health.