Add a key command which allows the player to manually attack
When the key is pressed the an anonymous function is created and
intercepts all keypresses until it is either aborted or gets some valid
input. In this case the player can select a direction in which he will
attack. Once he has picked a direction he can confirm it by pressing the
attack key again and the interface sends the command to the game logic
where it is performed.