Update the Interact Action class to use a direction while performing
Extended the selectTarget function to be re-usable for other actions
instead of just for the Attack action.

The player can now choose which doors to open and to close when standing
adjacent to multiple doors. Before this, the player would always open
and close the same door when there were multiple choices.