Add mechanics for simple ranged combat
The player can enter the selection "phase" (by hitting the F-Key) in
which he can use the directional keys to select a target to attack. We
use Bresenham's line algorithm to check if the target is in range,
visible and not blocked (for example by a wall).

If the player hits the F-Key again the target is sent from the interface
class to the game logic, where a RangedAttack action is created which
then calculates the outcome of the ranged attack.