Merge functions for selecting a target
All actions which require to select a target tile are now using the same
function for specifying that target. The player can move the cursor with
the arrow keys and the function will use Bresenham's line algorithm to
check if there is an unblocked path from the origin to the target tile.

It takes into account if the tile lies within range of the origin, if
there are any blocking tiles in the way and if the tile lies outside of
the game's grid.

The cursor can never be moved farther away from the origin than
specified by the calling action (e.g. the actions for opening doors only
have a range of one, which means the player has to be directly adjacent
to a door to select it as a target).