Set foundation for having equipable items
Moved the inventory related interface to the InputHandler for now (it
should be moved to its own class later on). The player can switch to
the inventory by pressing the I-Key. From there he or she can select
which item to equip by using the up and down keys and confirming the
selection with the E-Key.

From there a message is sent to the game logic where a new Equip action
will be created for the player and the selected item. It will move the
item to the equipment slot and remove it from the main inventory.