Add default weapons for each actor
Each actor gets a default weapon when it is created. For dogs and cats
this is the "Claw" item for example. This default item will always be
used when an actor unequips a weapon.

This also fixes a crash, which was caused when the player tried to
attack an enemy without a weapon in his or her hand.