Use "dirty" flags to mark tiles which need to be updated in the batch
When a tile changes its state we set it to dirty. This tells the ui that
it might need to update the graphical state of the tile as well and
causes it to send the updated information about the tile to the
spritebatch.

At the beginning of each turn, visible tiles are set to "not visible"
and their dirty flag is set to true. The same thing happens for tiles
which are seen by the player. This means only visible tiles will be
updated each turn which keeps the calculation cost low and the
performance high.

A positive side effect of this is, that the if there is a door which is
outside of the visible area and is opened by another actor won't show up
as open until it appears in the visible area of the player.