Add a random dungeon generator
Using a BSP tree to split the map into smaller partitions and then spawn
rooms randomly across these partitions. Rooms are finally connected via
corridors leading from the center of one room to the center of the
other.

For testing purposes the map is saved to a text file after it has been
created. This way it is easier to see changes without having to run
through the dungeon in the game itself.