Add simple pathfinding from NPCs to the player character
First we check if the actor can see the player by making a LOS check
from the NPC to the player using bresenham's line algorithm.

When the npc sees the player we set the player's position as the new
target. This target is then used by the NPC to calculate the movement
action for the next turn.

The action system makes sure that the NPCs open doors and attack the
player.