Add ItemFactory which takes care of loading templates and creating items
All items are now defined as template files in the resources folder. The
ItemFactory loads all files from the given paths, runs them as lua files
(to make them return the table containing the item definition) and
stores them in a table.

When we want to create an item in the game we call the public createItem
function and the ItemFactory will create it based on the template file.